To avoid 64bit build error

+++ src/player.c
@@ -97,7 +97,7 @@
 			{ \
 				g_cond_wait (&handle->message_queue_cond,&handle->message_queue_lock); \
 			} \
-			handle->current_message = (int)g_queue_pop_head (handle->message_queue); \
+			handle->current_message = (int)(intptr_t)g_queue_pop_head (handle->message_queue); \
 			g_mutex_unlock (&handle->message_queue_lock); \
 			LOGI("Retrived  message [%d] from queue",handle->current_message); \
 		}else{ \
